requirements_ubuntu.txt
actionlib==1.14.0
alabaster==0.7.13
ale-py==0.7.1
anaconda==0.0.1.1
anaconda-client==1.2.2
angles==1.9.13
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
arviz==0.11.4
astroid==2.11.7
astropy==4.1
asttokens==2.2.1
async-generator==1.10
atomicwrites==1.4.1
attrs==22.2.0
Automat==22.10.0
autopep8==2.0.2
Babel==2.11.0
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.tempfile==1.0
backports.weakref==1.0.post1
base_local_planner==1.17.2
bcrypt==4.0.1
beautifulsoup4==4.11.2
binaryornot==0.4.4
bitarray==2.7.3
bkcharts==0.2
black==22.8.0
bleach==4.1.0
blosc==1.10.6
bokeh==2.3.3
boto3==1.23.10
botocore==1.26.10
Bottleneck==1.3.7
box2d-py==2.3.5
Brotli==1.0.9
brotlipy==0.7.0
cached-property==1.5.2
camera-calibration-parsers==1.12.0
catkin==0.8.10
certifi==2021.5.30
cffi==1.15.1
cftime==1.6.0
chardet==5.0.0
charset-normalizer==2.0.12
click==8.0.4
cloudpickle==1.2.2
clyent==1.2.1
colorama==0.4.5
colorcet==3.0.1
comtypes==1.1.14
conda==4.3.16
conda-build==2.1.5
conda-content-trust==0.1.2
conda-env==2.4.2
conda-pack==0.6.0
constantly==15.1.0
contextvars==2.4
controller-manager==0.19.6
controller-manager-msgs==0.19.6
cookiecutter==1.7.3
cryptography==39.0.2
cssselect==1.1.0
cv-bridge==1.16.2
cycler==0.11.0
Cython==0.29.33
cytoolz==0.12.1
daal==2021.5.3
daal4py==2021.5.3
dask==2021.3.0
dataclasses==0.8
datashader==0.13.0
datashape==0.5.2
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diagnostic-updater==1.11.0
diff-match-patch==20200713
docutils==0.18.1
dynamic-reconfigure==1.7.3
entrypoints==0.4
et-xmlfile==1.1.0
executing==1.2.0
fasteners==0.18
filelock==3.4.1
flake8==2.5.5
Flask==2.0.3
fonttools==4.27.1
fsspec==2022.1.0
future==0.18.3
gazebo_plugins==2.9.2
gazebo_ros==2.9.2
gencpp==0.7.0
geneus==3.0.0
genlisp==0.4.18
genmsg==0.6.0
gennodejs==2.0.2
genpy==0.6.15
gensim==4.2.0
glfw==2.5.6
glob2==0.7
greenlet==2.0.2
gym==0.15.6
h5py==3.1.0
HeapDict==1.0.1
holoviews==1.14.9
hvplot==0.8.2
hyperlink==21.0.0
icc-rt==2020.0.133
icu==0.0.1
idna==3.4
image-geometry==1.16.2
imagecodecs==2020.5.30
imageio==2.15.0
imagesize==1.4.1
immutables==0.19
importlib-metadata==4.8.3
importlib-resources==5.4.0
incremental==22.10.0
inflection==0.5.1
iniconfig==1.1.1
intake==0.6.1
intel-openmp==2020.0.133
interactive-markers==1.12.0
intervaltree==3.1.0
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
ipywidgets==7.7.3
isort==5.10.1
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.17.2
jeepney==0.7.1
Jinja2==3.0.3
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.1
joint-state-publisher==1.15.1
joint-state-publisher-gui==1.15.1
joint_trajectory_action_tools==0.0.12
jq==1.4.0
json5==0.9.11
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.1.2
jupyter-console==6.4.3
jupyter-core==4.9.2
jupyter-server==1.13.1
jupyterlab==3.2.9
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.1.2
keyring==23.4.1
kiwisolver==1.3.1
laser_geometry==1.6.7
lazy-object-proxy==1.7.1
lerc==0.1.0
llvmlite==0.36.0
locket==1.0.0
lockfile==0.12.2
lxml==4.9.2
lz4==3.1.10
Markdown==3.3.7
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.4.0
message-filters==1.15.15
mistune==0.8.4
moveit-commander==1.1.11
moveit-core==1.1.11
moveit-ros-planning-interface==1.1.11
moveit-ros-visualization==1.1.11
msgpack==1.0.5
mujoco-py==1.50.1.68
multipledispatch==0.6.0
mypy-extensions==1.0.0
nbclassic==0.3.5
nbclient==0.5.9
nbconvert==6.0.7
nbformat==5.1.3
nest-asyncio==1.5.6
netCDF4==1.6.2
notebook==6.4.10
numba==0.53.1
numpy==1.19.5
opencv-python==4.7.0.72
packaging==21.3
pandas==1.1.5
pandocfilters==1.5.0
panel==0.12.1
param==1.12.3
parsel==1.6.0
parso==0.7.1
partd==1.2.0
pathspec==0.9.0
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.4.0
platformdirs==2.4.0
pluggy==1.0.0
poyo==0.5.0
pr2_controller_manager==1.8.21
prometheus-client==0.16.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
py==1.11.0
pycodestyle==2.10.0
pycosat==0.6.3
pycparser==2.21
pyct==0.4.8
pyflakes==1.0.0
pyglet==1.5.0
Pygments==2.14.0
pyparsing==3.0.9
pyrsistent==0.18.0
pytest==7.0.1
python-dateutil==2.8.2
python-qt-binding==0.4.4
python-slugify==6.1.2
pytz==2022.7.1
pyviz-comms==2.2.1
PyYAML==6.0
pyzmq==25.0.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-py-common==0.4.2
qtconsole==5.2.2
QtPy==2.0.1
requests==2.27.1
resource_retriever==1.12.7
rosapi==0.11.16
rosbag==1.15.15
rosboost-cfg==1.15.8
rosbridge-library==0.11.16
rosbridge-server==0.11.16
rosclean==1.15.8
rosgraph==1.15.15
roslaunch==1.15.15
roslib==1.15.8
roslint==0.12.0
roslz4==1.15.15
rosmake==1.15.8
rosmaster==1.15.15
rosmsg==1.15.15
rosnode==1.15.15
rosparam==1.15.15
rospy==1.15.15
rospy_message_converter==0.5.9
rosserial_python==0.9.2
rosservice==1.15.15
rostest==1.15.15
rostopic==1.15.15
rosunit==1.15.8
roswtf==1.15.15
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
rviz==1.14.19
s3transfer==0.5.2
scipy==1.5.4
SecretStorage==3.3.3
Send2Trash==1.8.0
sensor-msgs==1.13.1
six==1.16.0
smach==2.5.0
smart-open==6.3.0
smclib==1.8.6
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
srdfdom==0.6.3
tbb==2021.8.0
tblib==1.7.0
terminado==0.12.1
testpath==0.6.0
text-unidecode==1.3
tf==1.13.2
tf-conversions==1.13.2
tf2-geometry-msgs==0.7.6
tf2-kdl==0.7.6
tf2-py==0.7.6
tf2-ros==0.7.6
tomli==1.2.3
toolz==0.12.0
topic-tools==1.15.15
tornado==6.1
tqdm==4.64.1
traitlets==4.3.3
typed-ast==1.5.4
typing-extensions==3.10.0.2
urdfdom-py==0.4.6
urllib3==1.26.14
w3lib==2.0.1
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.3.1
Werkzeug==2.0.3
widgetsnbextension==3.6.2
wrapt==1.15.0
xacro==1.14.14
xarray==0.16.2
zict==2.1.0
zipp==3.6.0
